Update the jobs every 1 minute
